requirements.txt
Django==2.2.1
djangorestframework==3.9.4
gitdb2==2.0.5